extract hooks tests to separate file
the client tests are in a separate package,
and the hook-chaining tests aren't relevant
for generic clients.
